Add files via upload
This is a node red flow that compares the current time against a set time and sends a message if there is a match. I plan to use this to automatically start the Hot Liquor tank to heat up so it is ready to brew first thing in the morning. There is also a activation switch and a indicator in the flow.